add strategy.single.nianhe.py

+++ sh/rm_desk_jpg.sh
@@ -1,3 +1,12 @@
 
-rm ../*.jpg
-rm ../*.png
+jpg_size=$(ls ..|grep [.]jpg|wc -l)
+if [ $jpg_size -gt 0 ]
+then
+	rm ../*.jpg
+fi
+
+png_size=$(ls ..|grep [.]png|wc -l)
+if [ $png_size -gt 0 ]
+then
+	rm ../*.png
+fi
